Move template stuff around
